scanpos-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
marshmallow==3.20.1
python-dotenv==1.0.0
